Split the event.rs into a full module.
